% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon.R
\name{neon_product}
\alias{neon_product}
\title{Fetch details for a specific NEON data product}
\usage{
neon_product(code)
}
\arguments{
\item{code}{NEON product code (e.g. "DP1.10003.001")}
}
\value{
tibble: one row with productCode, productName, productDescription,
productScienceTeam, productStatus, siteCodes
}
\description{
Fetch details for a specific NEON data product
}
