% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon.R
\name{neon_products}
\alias{neon_products}
\title{List all NEON data products}
\usage{
neon_products()
}
\value{
tibble: productCode, productName, productDescription,
productScienceTeam, productStatus
}
\description{
List all NEON data products
}
