% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon.R
\name{neon_sites}
\alias{neon_sites}
\title{List all NEON field sites}
\usage{
neon_sites()
}
\value{
tibble: siteCode, siteDescription, siteType, stateName,
domainCode, siteLatitude, siteLongitude
}
\description{
List all NEON field sites
}
